Make User Data Downloadable and Saveable!
The user data was stored in a new repository that Spring Boot put in, then React downloads 75 random users, and those users get sent and updated into the User repository. Then, now that the data is stored inside the User repository, you can use a GET request and download ALL of the User data and parse it to Java Object, and save it using the Serializable Interface!!!!! Finally, data can get reloaded from the saved file and have fun. :)